+++ src/main/groovy/com/exictos/devops/containers/WebSphere.groovy
@@ -18,7 +18,7 @@ class WebSphere extends Container{
     WebSphere(String host, int port, String username, char[] password, String aWsadmin)
     {
         wsadmin = new WSAdminWrapper(aWsadmin,host,port,username,password)
-        profile = new WebSphereProfile(wsadmin)
+        profile = new WebSphereProfile(wsadmin, log)
         profile.host = host
         profile.port = port
         profile.username = username
@@ -54,15 +54,15 @@ class WebSphere extends Container{
     protected String installApp(File pathToPackage, String applicationName, String applicationVersion
                                 , Timestamp timestamp)
     {
-        log.info("Installing application ${applicationName} from package at ${pathToPackage}...")
+        log.log("Installing application ${applicationName} from package at ${pathToPackage}...")
         try{
             String name = LiberoHelper.standardizeName(pathToPackage.getAbsolutePath(), applicationName, applicationVersion
                     , timestamp)
             wsadmin.installApplication(pathToPackage.getAbsolutePath(),name)
-            log.info("${applicationName} installed successfully as ${name}")
+            log.log("${applicationName} installed successfully as ${name}")
             return name
         }catch(Exception e){
-            log.error "Could not install application ${applicationName} from package ${pathToPackage}. Cause: ${e.getMessage()}"
+            log.log "Could not install application ${applicationName} from package ${pathToPackage}. Cause: ${e.getMessage()}"
             throw e
         }
     }
@@ -75,13 +75,13 @@ class WebSphere extends Container{
     @Override
     void startApp(String deploymentName)
     {
-        log.info("Starting application: ${deploymentName}...")
+        log.log("Starting application: ${deploymentName}...")
         try{
             if(wsadmin.isAppReady(deploymentName))
                 wsadmin.startApplication(deploymentName)
-            log.info("Deployment ${deploymentName} started.")
+            log.log("Deployment ${deploymentName} started.")
         }catch(Exception e){
-            log.error("Could not start application ${deploymentName}. Cause: ${e}")
+            log.log("Could not start application ${deploymentName}. Cause: ${e}")
             throw e
         }
     }
@@ -95,12 +95,12 @@ class WebSphere extends Container{
     @Override
     void stopApp(String deploymentName)
     {
-        log.info("Stopping deployment ${deploymentName}...")
+        log.log("Stopping deployment ${deploymentName}...")
         try{
             wsadmin.stopApplication(deploymentName)
-            log.info("Deployment ${deploymentName} stopped")
+            log.log("Deployment ${deploymentName} stopped")
         }catch(Exception e){
-            log.error("Could not stop deployment: ${deploymentName}. Cause: ${e}")
+            log.log("Could not stop deployment: ${deploymentName}. Cause: ${e}")
             throw e
         }
     }
@@ -113,12 +113,12 @@ class WebSphere extends Container{
     @Override
     void uninstallApp(String deploymentName)
     {
-        log.info("Uninstalling deployment ${deploymentName}...")
+        log.log("Uninstalling deployment ${deploymentName}...")
         try{
             wsadmin.uninstallApp(deploymentName)
-            log.info("Deployment ${deploymentName} uninstalled.")
+            log.log("Deployment ${deploymentName} uninstalled.")
         }catch(Exception e){
-            log.error "Could not uninstall ${deploymentName}. Cause: ${e}"
+            log.log "Could not uninstall ${deploymentName}. Cause: ${e}"
             throw e
         }
     }